requirements.txt
# Image handling for Tkinter GUI
Pillow>=9.0,<11.0

# Serial port communication
pyserial>=3.5,<4.0


# Report Generation in Excel 
openpyxl 

# Report Generation in PDF Form

reportlab 